Fixing last two commits.  Eclipse did not understand that refactoring.
